+++ packages/ui/src/functions/keyboard.ts
@@ -64,13 +64,17 @@ const KEY_FN: Record<string, (player: Player) => void> = {
 export default function (it: UIInterface) {
   const { player, config } = it
 
+  if (typeof config.keyboard == 'undefined') {
+    config.keyboard = { focused: true }
+  }
+
   function keydown(e: KeyboardEvent) {
     if (
       document.activeElement?.tagName == 'INPUT' ||
       document.activeElement?.tagName == 'TEXTAREA' ||
       document.activeElement?.getAttribute('contenteditable') ||
-      (!config.keyboard?.global && !config.keyboard?.focused) ||
-      (config.keyboard.focused && !isFocused(player)) ||
+      (!config.keyboard!.global && !config.keyboard!.focused) ||
+      (config.keyboard!.focused && !isFocused(player)) ||
       e.altKey ||
       e.ctrlKey ||
       e.metaKey ||