+++ packages/ui/src/index.ts
@@ -33,7 +33,6 @@ const defaultConfig: UiConfig = {
   forceLandscapeOnFullscreen: true,
 
   showControls: 'always',
-  keyboard: { focused: true },
   settings: ['loop'],
   theme: { primaryColor: '#6668ab' },
   speeds: ['2.0', '1.5', '1.25', '1.0', '0.75', '0.5'],